Move priceinput validator to handyFunctions for re-use
